Revert "Moved get_glyphsets_fulfilled into conditions"
This reverts commit b572a8f2d36699289974c5cb0b7cd2391184f502.
